--- a/PaddleNLP/language_model/args.py
+++ b/PaddleNLP/language_model/args.py
@@ -20,6 +20,15 @@ import argparse
 import distutils.util
 
 
+def str2bool(v):
+    if v.lower() in ('yes', 'true', 't', 'y', '1'):
+        return True
+    elif v.lower() in ('no', 'false', 'f', 'n', '0'):
+        return False
+    else:
+        raise argparse.ArgumentTypeError('Unsupported value encountered.')
+
+
 def parse_args():
     parser = argparse.ArgumentParser(description=__doc__)
     parser.add_argument(
@@ -36,10 +45,25 @@ def parse_args():
         "--data_path", type=str, help="all the data for train,valid,test")
     parser.add_argument('--para_init', action='store_true')
     parser.add_argument(
-        '--use_gpu', type=bool, default=False, help='whether using gpu')
+        '--use_gpu',
+        type=str2bool,
+        default=False,
+        help='Whether using gpu [True|False]')
+    parser.add_argument(
+        '--parallel',
+        type=str2bool,
+        default=True,
+        help='Whether using gpu in parallel [True|False]')
+    parser.add_argument(
+        '--use_py_reader',
+        type=str2bool,
+        default=False,
+        help='Whether using py_reader to feed data [True|False]')
     parser.add_argument(
         '--log_path',
         help='path of the log file. If not set, logs are printed to console')
     parser.add_argument('--enable_ce', action='store_true')
+    parser.add_argument('--batch_size', type=int, default=0, help='batch size')
+    parser.add_argument('--max_epoch', type=int, default=0, help='max epoch')
     args = parser.parse_args()
     return args